fix : menuPrice 자료형 변경

+++ src/main/java/com/example/yamyam16/menu/dto/MenuCreateRequestDto.java
@@ -15,5 +15,5 @@ public class MenuCreateRequestDto { //requestBody로 오는 내용만 적음
 
 	@NotNull(message = "메뉴 가격은 필수입니다.")
 	@Positive(message = "메뉴 가격은 0보다 커야 합니다")
-	private Integer menuPrice; //@notnull 검증할 때는 integer을 사용, dto에서 자주 사용한다. int는 단순 내부 계산할 때만 사용
+	private Long menuPrice; //@notnull 검증할 때는 integer을 사용, dto에서 자주 사용한다. int는 단순 내부 계산할 때만 사용
 }
